changed nemesys structure
